Extract getEmailParams into function
